utils: modify type of value

+++ include/utils.h
@@ -43,7 +43,7 @@ float int16_to_signed_dec(unsigned short value, int range);
  *  @param range is the range of the value.
  *  @return the encoded value.
  */
-int signed_dec_to_int32(float value, int range);
+int signed_dec_to_int32(double value, int range);
 
 /**
  *  Convert an int32 value to a signed decimal.
@@ -53,7 +53,7 @@ int signed_dec_to_int32(float value, int range);
  *  @param range is the range of the value.
  *  @return the decoded value.
  */
-float int32_to_signed_dec(int value, int range);
+double int32_to_signed_dec(int value, int range);
 
 
 #endif /* !UTILS_H */
